+++ smtlib-backends-process/CHANGELOG.md
@@ -8,6 +8,8 @@
     `Process.close`, hence gracefully
 - add a `Process.write` function for writing commands without reading the
   solver's response
+- add a test checking that we can pile up procedures for exiting a process
+  safely
 
 # v0.2
 split `smtlib-backends`'s `Process` module into its own library