+++ smtlib-backends-process/CHANGELOG.md
@@ -11,6 +11,7 @@
 - add a test checking that we can pile up procedures for exiting a process
   safely
 - add `Process.defaultConfig`, synonym for `def`
+- improve error messages inside `Process.toBackend`
 
 # v0.2
 split `smtlib-backends`'s `Process` module into its own library