+++ smtlib-backends-process/CHANGELOG.md
@@ -6,8 +6,6 @@
   - send an `(exit)` command before waiting for the process to exit
   - this means `Process.with` now closes the process with this new version of
     `Process.close`, hence gracefully
-- add a `Process.write` function for writing commands without reading the
-  solver's response
 - add a test checking that we can pile up procedures for exiting a process
   safely
 - add `Process.defaultConfig`, synonym for `def`
@@ -19,7 +17,7 @@ split `smtlib-backends`'s `Process` module into its own library
 - move the logger function into it
 - make it an instance of the `Default` typeclass
 ## logging
-- move the logger function into the `Config` datatype 
+- move the logger function into the `Config` datatype
 - don't prefix error messages with `[stderr]`
 ## test-suite
 - add usage examples